requirements.txt
pyserial
opencv-python
Pillow
customtkinter
numpy